Started refactoring of beta
